+++ src/Libplanet.Net/Messages/GetBlockHashesMsg.cs
@@ -1,6 +1,5 @@
 using System;
 using System.Collections.Generic;
-using System.Linq;
 using Libplanet.Blockchain;
 using Libplanet.Types.Blocks;
 
@@ -8,26 +7,17 @@ namespace Libplanet.Net.Messages
 {
     internal class GetBlockHashesMsg : MessageContent
     {
-        public GetBlockHashesMsg(BlockLocator locator, BlockHash? stop)
+        public GetBlockHashesMsg(BlockLocator locator)
         {
             Locator = locator;
-            Stop = stop;
         }
 
         public GetBlockHashesMsg(byte[][] dataFrames)
         {
-            int requestedHashCount = BitConverter.ToInt32(dataFrames[0], 0);
-            Locator = new BlockLocator(
-                dataFrames.Skip(1).Take(requestedHashCount)
-                .Select(frame => new BlockHash(frame)));
-            Stop = dataFrames[1 + requestedHashCount].Length == 0
-                ? default(BlockHash?)
-                : new BlockHash(dataFrames[1 + requestedHashCount]);
+            Locator = new BlockLocator(new BlockHash(dataFrames[1]));
         }
 
         public BlockLocator Locator { get; }
-
-        public BlockHash? Stop { get; }
 
         public override MessageType Type => MessageType.GetBlockHashes;
 
@@ -36,11 +26,9 @@ namespace Libplanet.Net.Messages
             get
             {
                 var frames = new List<byte[]>();
-                frames.Add(BitConverter.GetBytes(Locator.Count()));
-                frames.AddRange(Locator.Select(hash => hash.ToByteArray()));
-                frames.Add(Stop is { } s
-                    ? s.ToByteArray()
-                    : new byte[] { });
+                frames.Add(BitConverter.GetBytes(1));
+                frames.Add(Locator.Hash.ToByteArray());
+                frames.Add(Array.Empty<byte>());
                 return frames;
             }
         }